refactor(esl-share): refactor ESLShareConfig
