+++ src/modules/esl-share/core/esl-share.ts
@@ -7,7 +7,7 @@ import {ESLShareTrigger} from './esl-share-trigger';
 import {ESLShareConfig} from './esl-share-config';
 
 import type {PopupActionParams} from '../../esl-popup/core';
-import type {ESLShareConfigShape, ESLShareButtonConfig, ESLShareConfigButtons, ESLShareConfigProviderType} from './esl-share-config';
+import type {ESLShareConfigShape, ESLShareButtonConfig, ESLShareConfigProviderType} from './esl-share-config';
 
 /**
  * ESLShare
@@ -28,9 +28,9 @@ export class ESLShare extends ESLBaseElement {
 
   /**
    * Gets or updates config with a promise of a new config object or using a config provider function.
-   * @returns Promise of the current config {@link ESLShareConfigButtons}
+   * @returns Promise of the current config {@link ESLShareConfig}
    */
-  public static config(provider?: ESLShareConfigProviderType | ESLShareConfigShape): Promise<ESLShareConfigButtons> {
+  public static config(provider?: ESLShareConfigProviderType | ESLShareConfigShape): Promise<ESLShareConfig> {
     return ESLShareConfig.set(provider);
   }
